# Conflicts:
#	src/Orleans/Configuration/GlobalConfiguration.cs
#	src/Orleans/Logging/ErrorCodes.cs
#	src/Orleans/Orleans.csproj
#	src/Orleans/Providers/IProviderRuntime.cs
#	src/OrleansRuntime/Catalog/Catalog.cs
#	src/OrleansRuntime/GrainTypeManager/GrainTypeData.cs
#	src/OrleansRuntime/GrainTypeManager/SiloAssemblyLoader.cs
#	src/OrleansRuntime/MultiClusterNetwork/MultiClusterOracleData.cs
#	src/OrleansRuntime/OrleansRuntime.csproj
#	src/OrleansRuntime/Silo/Silo.cs
#	src/OrleansRuntime/Storage/MemoryStorageGrain.cs
#	src/OrleansTestingHost/OrleansTestingHost.csproj
#	test/TestGrains/TestGrains.csproj
#	test/TesterInternal/GeoClusterTests/TestingClusterHost.cs
#	test/TesterInternal/TesterInternal.csproj

+++ src/Orleans/Streams/SimpleMessageStream/SimpleMessageStreamProducerExtension.cs
@@ -1,9 +1,7 @@
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
-
 using Orleans.Concurrency;
 using Orleans.Runtime;
 using Orleans.Streams;
@@ -227,7 +225,7 @@ namespace Orleans.Providers.Streams.SimpleMessageStream
                     if (consumers.TryRemove(subscriptionId, out discard))
                     {
                         streamPubSub.UnregisterConsumer(subscriptionId, streamId, streamId.ProviderName).Ignore();
-                        logger.Warn((int)ErrorCode.Stream_ConsumerIsDead,
+                        logger.Warn(ErrorCode.Stream_ConsumerIsDead,
                             "Consumer {0} on stream {1} is no longer active - permanently removing Consumer.", remoteConsumer, streamId);
                     }
                 }
